pyproject.toml
[project]
name = "korean-voice-clone-tts"
version = "0.1.0"
description = "Korean-first voice cloning TTS service"
requires-python = ">=3.13"
dependencies = [
    # Core
    "torch>=2.0",
    "torchaudio>=2.0",
    "fastapi==0.134.0",
    "uvicorn>=0.20",
    "celery>=5.3",
    "redis>=5.0",
    "pydantic-settings>=2.0",
    # Audio Processing
    "demucs>=4.0",
    "deepfilternet>=0.5",
    "pyloudnorm>=0.1",
    "soxr>=0.3",
    "soundfile>=0.12",
    # Speech Processing
    "pyannote-audio>=3.1",
    "faster-whisper>=1.0",
    "speechbrain>=1.0",
    # Korean NLP
    "g2pk>=0.9",
    "python-mecab-ko>=1.0",
    "jamo>=0.4",
    # Storage
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "python-multipart>=0.0.22",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "httpx>=0.24",
    "ruff>=0.4",
    "ty>=0.0.1a1",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]
